M02A02 – Coerção de Tipo

+++ Modulo_02_Operadores/M02A02–Coercao_de_Tipo.c
@@ -2,11 +2,11 @@
 #include <locale.h>
 
 void main(){
-setlocale(LC_ALL,"Portuguese");
+setlocale(LC_ALL, "Portuguese");
 
-    float x = 5.0;
+    int x = 5.0;
     int y = 2;
-    float z = x / y;
+    float z = (float)x / y;
 
     printf("O resultado e %.1f",z);
 }
